ADDED: UC-2

+++ src/com/bridgelabz/Utility/GenericClass.java
@@ -3,13 +3,14 @@ package com.bridgelabz.Utility;
 public class GenericClass {
     /**
      * A method to find the maximum among three integers
-     * @param x First Input Integer
-     * @param y Second Input Integer
-     * @param z Third Input Integer
-     * @return max Max is the greatest among all entered integers
+     * @param x First Input Double
+     * @param y Second Input Double
+     * @param z Third Input Double
+     * @return max Max is the greatest among all entered numbers
      */
-    public static Integer getMax( Integer x,Integer y,Integer z){
-        Integer max=x;
+
+    public static Double getMax( Double x,Double y,Double z){
+        Double max=x;
         if(y.compareTo(max)>0){
             max=y;
         }
@@ -18,6 +19,5 @@ public class GenericClass {
         }
         return max;
     }
-
 
 }
